requirements.txt
aiohttp==3.8.1
python-telegram-bot==20.0
